+++ programs/tributary/src/policies/mod.rs
@@ -10,8 +10,10 @@ pub mod milestone;
 pub mod one_time;
 pub mod pay_as_you_go;
 pub mod subscription;
+pub mod up_to;
 
 pub use milestone::validate_milestone_policy;
 pub use one_time::validate_one_time_policy;
 pub use pay_as_you_go::validate_payg_policy;
 pub use subscription::validate_subscription_policy;
+pub use up_to::validate_up_to_policy;